+++ src/_view/handling.c
@@ -32,6 +32,22 @@
 #define INITIAL_BUF_SIZE 1024
 
 /*
+ * Print an error without clearing it.
+ */
+static void
+show_error(bool dev)
+{
+    if (dev)
+    {
+        PyObject *err = PyErr_GetRaisedException();
+        Py_INCREF(err); // Save a reference to it
+        PyErr_SetRaisedException(err);
+        PyErr_Print();
+        PyErr_SetRaisedException(err);
+    } else PyErr_Clear();
+}
+
+/*
  * Call a route object with both query and body parameters.
  */
 int
@@ -85,7 +101,7 @@ handle_route_impl(
 
     if (!query_obj)
     {
-        PyErr_Clear();
+        show_error(self->dev);
         return server_err(
             self,
             awaitable,
@@ -112,11 +128,7 @@ handle_route_impl(
 
     if (!params)
     {
-        // parsing failed
-        if (self->dev)
-            PyErr_Print();
-        else PyErr_Clear();
-
+        show_error(self->dev);
         return server_err(
             self,
             awaitable,