+++ src/_view/inputs.c
@@ -4,11 +4,18 @@
  * This file is responsible for parsing route inputs through query
  * strings and body parameters.
  *
- * These functions are only called when there inputs on a route - otherwise the
- * parsing is skipped.
+ * If a route has no inputs, then the parsing
+ * step is skipped for optimization purposes.
+ *
+ * If a route has only query inputs, then we don't need to go through the
+ * body parsing step, and only parse the query string (handle_route_query()).
+ *
+ * If a route has body inputs, then we start by parsing that, and if it has any
+ * query string parameters, that's handled later. ASGI does not send the body
+ * in a single receive() call, so we have a buffer that increases over time.
  *
  * This implementation is also in charge of building data inputs (such as Context() or WebSocket())
- * and appending them to routes.
+ * and appending them to routes. This is indicated by a special integer determined by the loader.
  *
  */
 #include <Python.h>
@@ -187,7 +194,6 @@ body_inc_buf(PyObject *awaitable, PyObject *result)
             PyMem_Free(nbuf);
             return -1;
         }
-        ;
     }
 
     Py_DECREF(more_body);
@@ -448,6 +454,13 @@ query_parser(
     return obj; // no need for null check
 }
 
+/*
+ * Build a route data object based on the given data ID (determined by the loader).
+ *
+ * As of now:
+ * - 1: Context()
+ * - 2: WebSocket(), only supported on WebSocket routes
+ */
 PyObject *
 build_data_input(
     int num,
@@ -587,5 +600,7 @@ generate_params(
         ob[i] = item;
     }
 
+    // TODO: fix this reference leak
+    //Py_DECREF(obj);
     return ob;
 }