+++ CHANGELOG.md
@@ -35,8 +35,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added the `view dev` command (live reload)
 -   Fixed redirection and disabling of HTTP server logging
 -   C API is now compliant with [PEP 7](https://peps.python.org/pep-0007/)
--   Added `-g3` flag to the `_view` extension module (debugging information)
+-   Added `-g3` and `-O3` flag to the `_view` extension module (debugging information and optimizations)
 -   Removed use of `escape()` in the message shown when a dependency is needed
+-   Query string client errors are now displayed during development mode
+-   `KeyboardInterrupt` is swallowed by the server coroutine, and a log message is now issued
 -   **Breaking Change:** Renamed `Error` to `HTTPError`
 -   **Breaking Change:** `__view_result__` is now given a `Context` parameter
 -   **Breaking Change:** `to_response` is now asynchronous